+++ src/ManagedTypes/DoubleManagedType.cs
@@ -20,7 +20,7 @@ internal class DoubleManagedType : IManagedType
     /// <summary>
     /// Returns the value between start and end for the percentage passed in.
     /// </summary>
-    public object GetIntermediateValue(object start, object end, double percentage) => Utility.Interpolate((double)start, (double)end, percentage);
+    public object GetIntermediateValue(object start, object end, float percentage) => Utility.Interpolate((double)start, (double)end, percentage);
 
     #endregion
 }
