+++ src/ManagedTypes/IManagedType.cs
@@ -24,6 +24,6 @@ internal interface IManagedType
     /// to the percentage passed in. (Note: the percentage can be less than 0% or
     /// greater than 100%.)
     /// </summary>
-    object GetIntermediateValue(object start, object end, double percentage);
+    object GetIntermediateValue(object start, object end, float percentage);
 
 }
